Add String-based Gesture Extensions
