+++ src/CommunityToolkit.Maui.Markup/GesturesExtensions.cs
@@ -1,256 +1,53 @@
-﻿using System.Linq.Expressions;
+﻿using System.Diagnostics.CodeAnalysis;
 using System.Windows.Input;
+
 namespace CommunityToolkit.Maui.Markup;
 
 /// <summary>
 /// Extension Methods for Element Gestures
 /// </summary>
-public static class GesturesExtensions
+public static partial class GesturesExtensions
 {
-	/// <summary>Add a <see cref="SwipeGestureRecognizer"/> and bind to its Command </summary>
-	public static TGestureElement BindSwipeGesture<TGestureElement, TCommandBindingContext>(
-		this TGestureElement gestureElement,
-		Expression<Func<TCommandBindingContext, ICommand>> getter,
-		Action<TCommandBindingContext, ICommand>? setter = null,
-		TCommandBindingContext? source = default,
-		BindingMode mode = BindingMode.Default,
-		SwipeDirection? direction = null,
-		uint? threshold = null)
-		where TGestureElement : BindableObject, IGestureRecognizers
-		where TCommandBindingContext : class
-	{
-		return BindSwipeGesture<TGestureElement, TCommandBindingContext, object, object?>(
-			gestureElement,
-			getter,
-			setter,
-			source,
-			mode,
-			direction: direction,
-			threshold: threshold);
-	}
-
 	/// <summary>Add a <see cref="SwipeGestureRecognizer"/> and bind to its Command and (optionally) CommandParameter properties</summary>
-	public static TGestureElement BindSwipeGesture<TGestureElement, TCommandBindingContext, TParameterBindingContext, TParameterSource>(
-		this TGestureElement gestureElement,
-		Expression<Func<TCommandBindingContext, ICommand>> getter,
-		Action<TCommandBindingContext, ICommand>? setter = null,
-		TCommandBindingContext? source = default,
-		BindingMode commandBindingMode = BindingMode.Default,
-		Expression<Func<TParameterBindingContext, TParameterSource>>? parameterGetter = null,
-		Action<TParameterBindingContext, TParameterSource>? parameterSetter = null,
-		BindingMode parameterBindingMode = BindingMode.Default,
-		TParameterBindingContext? parameterSource = default,
-		SwipeDirection? direction = null,
-		uint? threshold = null)
-		where TGestureElement : BindableObject, IGestureRecognizers
-		where TCommandBindingContext : class
-		where TParameterBindingContext : class
+	/// <param name="gestureElement">An <see cref="Element"/> implementing <see cref="IGestureRecognizers"/></param>
+	/// <param name="commandPath">Path to Command Binding</param>
+	/// <param name="commandSource">Binding source for Command Binding</param>
+	/// <param name="parameterPath">If not specified or null, no binding is created for the CommandParameter property</param>
+	/// <param name="parameterSource">Binding source for Command Binding</param>
+	/// <param name="direction">Swipe gesture direction</param>
+	/// <param name="threshold">Minimum swipe distance that will cause the gesture to be recognized</param>
+	/// <returns><paramref name="gestureElement"/></returns>
+	[RequiresUnreferencedCode("Using bindings with string paths is not trim safe. Use expression-based binding instead.")]
+	public static TGestureElement BindSwipeGesture<TGestureElement>(this TGestureElement gestureElement,
+																		string commandPath,
+																		object? commandSource = null,
+																		string? parameterPath = null,
+																		object? parameterSource = null,
+																		SwipeDirection? direction = null,
+																		uint? threshold = null) where TGestureElement : IGestureRecognizers
 	{
-		var getterFunc = ConvertExpressionToFunc(getter);
-		var parameterGetterFunc = parameterGetter switch
-		{
-			null => null,
-			_ => ConvertExpressionToFunc(parameterGetter)
-		};
+		var swipeGesture = gestureElement.BindGesture<TGestureElement, SwipeGestureRecognizer>(commandPath, commandSource, parameterPath, parameterSource);
 
-		(Func<TParameterBindingContext, object?>, string)[]? parameterGetterHandlers = parameterGetter switch
-		{
-			null => null,
-			_ => [(b => b, GetMemberName(parameterGetter))]
-		};
-
-		return BindSwipeGesture(
-			gestureElement,
-			getterFunc,
-			[
-				(b => b, GetMemberName(getter))
-			],
-			setter,
-			source,
-			commandBindingMode,
-			parameterGetterFunc,
-			parameterGetterHandlers,
-			parameterSetter,
-			parameterBindingMode,
-			parameterSource,
-			direction,
-			threshold);
-	}
-
-	/// <summary>Add a <see cref="SwipeGestureRecognizer"/> and bind to its Command</summary>
-	public static TGestureElement BindSwipeGesture<TGestureElement, TCommandBindingContext>(
-		this TGestureElement gestureElement,
-		Func<TCommandBindingContext, ICommand> getter,
-		(Func<TCommandBindingContext, object?>, string)[] handlers,
-		Action<TCommandBindingContext, ICommand>? setter = null,
-		TCommandBindingContext? source = default,
-		BindingMode mode = BindingMode.Default,
-		SwipeDirection? direction = null,
-		uint? threshold = null)
-		where TGestureElement : BindableObject, IGestureRecognizers
-		where TCommandBindingContext : class
-	{
-		return gestureElement.BindSwipeGesture<TGestureElement, TCommandBindingContext, object, object?>(
-			getter,
-			handlers,
-			setter,
-			source,
-			mode,
-			direction: direction,
-			threshold: threshold);
-	}
-
-	/// <summary>Add a <see cref="SwipeGestureRecognizer"/> and bind to its Command and (optionally) CommandParameter properties</summary>
-	public static TGestureElement BindSwipeGesture<TGestureElement, TCommandBindingContext, TParameterBindingContext, TParameterSource>(
-		this TGestureElement gestureElement,
-		Func<TCommandBindingContext, ICommand> getter,
-		(Func<TCommandBindingContext, object?>, string)[] handlers,
-		Action<TCommandBindingContext, ICommand>? setter = null,
-		TCommandBindingContext? source = default,
-		BindingMode commandBindingMode = BindingMode.Default,
-		Func<TParameterBindingContext, TParameterSource>? parameterGetter = null,
-		(Func<TParameterBindingContext, object?>, string)[]? parameterHandlers = null,
-		Action<TParameterBindingContext, TParameterSource>? parameterSetter = null,
-		BindingMode parameterBindingMode = BindingMode.Default,
-		TParameterBindingContext? parameterSource = default,
-		SwipeDirection? direction = null,
-		uint? threshold = null)
-		where TGestureElement : BindableObject, IGestureRecognizers
-		where TCommandBindingContext : class
-		where TParameterBindingContext : class
-	{
-		var clickGesture = gestureElement.BindGesture<SwipeGestureRecognizer, TGestureElement, TCommandBindingContext, TParameterBindingContext, TParameterSource>(
-			getter,
-			handlers,
-			setter,
-			source,
-			commandBindingMode,
-			parameterGetter,
-			parameterHandlers,
-			parameterSetter,
-			parameterBindingMode,
-			parameterSource);
-
-		return gestureElement.ConfigureSwipeGesture(clickGesture, direction, threshold);
-	}
-
-	/// <summary>Adds a <see cref="TapGestureRecognizer"/> and bind its Command</summary>
-	public static TGestureElement BindTapGesture<TGestureElement, TCommandBindingContext>(
-		this TGestureElement gestureElement,
-		Expression<Func<TCommandBindingContext, ICommand>> getter,
-		Action<TCommandBindingContext, ICommand>? setter = null,
-		TCommandBindingContext? source = default,
-		BindingMode mode = BindingMode.Default,
-		int? numberOfTapsRequired = null)
-		where TGestureElement : BindableObject, IGestureRecognizers
-		where TCommandBindingContext : class
-	{
-
-		return BindTapGesture<TGestureElement, TCommandBindingContext, object, object?>(
-			gestureElement,
-			getter,
-			setter,
-			source,
-			mode,
-			numberOfTapsRequired: numberOfTapsRequired);
+		return gestureElement.ConfigureSwipeGesture(swipeGesture, direction, threshold);
 	}
 
 	/// <summary>Adds a <see cref="TapGestureRecognizer"/> and bind its Command and (optionally) CommandParameter properties</summary>
-	public static TGestureElement BindTapGesture<TGestureElement, TCommandBindingContext, TParameterBindingContext, TParameterSource>(
-		this TGestureElement gestureElement,
-		Expression<Func<TCommandBindingContext, ICommand>> getter,
-		Action<TCommandBindingContext, ICommand>? setter = null,
-		TCommandBindingContext? source = default,
-		BindingMode commandBindingMode = BindingMode.Default,
-		Expression<Func<TParameterBindingContext, TParameterSource>>? parameterGetter = null,
-		Action<TParameterBindingContext, TParameterSource>? parameterSetter = null,
-		BindingMode parameterBindingMode = BindingMode.Default,
-		TParameterBindingContext? parameterSource = default,
-		int? numberOfTapsRequired = null)
-		where TGestureElement : BindableObject, IGestureRecognizers
-		where TCommandBindingContext : class
-		where TParameterBindingContext : class
+	/// <param name="gestureElement">An <see cref="Element"/> implementing <see cref="IGestureRecognizers"/></param>
+	/// <param name="commandPath">Path to Command Binding</param>
+	/// <param name="commandSource">Binding source for Command Binding</param>
+	/// <param name="parameterPath">If not specified or null, no binding is created for the CommandParameter property</param>
+	/// <param name="parameterSource">Binding source for Command Binding</param>
+	/// <param name="numberOfTapsRequired">Number of taps required to trigger the <see cref="ICommand"/></param>
+	/// <returns><paramref name="gestureElement"/></returns>
+	[RequiresUnreferencedCode("Using bindings with string paths is not trim safe. Use expression-based binding instead.")]
+	public static TGestureElement BindTapGesture<TGestureElement>(this TGestureElement gestureElement,
+																	string commandPath,
+																	object? commandSource = null,
+																	string? parameterPath = null,
+																	object? parameterSource = null,
+																	int? numberOfTapsRequired = null) where TGestureElement : IGestureRecognizers
 	{
-		var getterFunc = ConvertExpressionToFunc(getter);
-		var parameterGetterFunc = parameterGetter switch
-		{
-			null => null,
-			_ => ConvertExpressionToFunc(parameterGetter)
-		};
-
-		(Func<TParameterBindingContext, object?>, string)[]? parameterGetterHandlers = parameterGetter switch
-		{
-			null => null,
-			_ => [(b => b, GetMemberName(parameterGetter))]
-		};
-
-		return BindTapGesture(
-			gestureElement,
-			getterFunc,
-			[
-				(b => b, GetMemberName(getter))
-			],
-			setter,
-			source,
-			commandBindingMode,
-			parameterGetterFunc,
-			parameterGetterHandlers,
-			parameterSetter,
-			parameterBindingMode,
-			parameterSource,
-			numberOfTapsRequired);
-	}
-
-	/// <summary>Adds a <see cref="TapGestureRecognizer"/> and bind its Command </summary>
-	public static TGestureElement BindTapGesture<TGestureElement, TCommandBindingContext>(
-		this TGestureElement gestureElement,
-		Func<TCommandBindingContext, ICommand> getter,
-		(Func<TCommandBindingContext, object?>, string)[] handlers,
-		Action<TCommandBindingContext, ICommand>? setter = null,
-		TCommandBindingContext? source = default,
-		BindingMode mode = BindingMode.Default,
-		int? numberOfTapsRequired = null)
-		where TGestureElement : BindableObject, IGestureRecognizers
-		where TCommandBindingContext : class
-	{
-		return gestureElement.BindTapGesture<TGestureElement, TCommandBindingContext, object, object?>(
-			getter,
-			handlers,
-			setter,
-			source,
-			mode,
-			numberOfTapsRequired: numberOfTapsRequired);
-	}
-
-	/// <summary>Adds a <see cref="TapGestureRecognizer"/> and bind its Command and (optionally) CommandParameter properties</summary>
-	public static TGestureElement BindTapGesture<TGestureElement, TCommandBindingContext, TParameterBindingContext, TParameterSource>(
-		this TGestureElement gestureElement,
-		Func<TCommandBindingContext, ICommand> getter,
-		(Func<TCommandBindingContext, object?>, string)[] handlers,
-		Action<TCommandBindingContext, ICommand>? setter = null,
-		TCommandBindingContext? source = default,
-		BindingMode commandBindingMode = BindingMode.Default,
-		Func<TParameterBindingContext, TParameterSource>? parameterGetter = null,
-		(Func<TParameterBindingContext, object?>, string)[]? parameterHandlers = null,
-		Action<TParameterBindingContext, TParameterSource>? parameterSetter = null,
-		BindingMode parameterBindingMode = BindingMode.Default,
-		TParameterBindingContext? parameterSource = default,
-		int? numberOfTapsRequired = null)
-		where TGestureElement : BindableObject, IGestureRecognizers
-		where TCommandBindingContext : class
-		where TParameterBindingContext : class
-	{
-		var tapGesture = gestureElement.BindGesture<TapGestureRecognizer, TGestureElement, TCommandBindingContext, TParameterBindingContext, TParameterSource>(
-			getter,
-			handlers,
-			setter,
-			source,
-			commandBindingMode,
-			parameterGetter,
-			parameterHandlers,
-			parameterSetter,
-			parameterBindingMode,
-			parameterSource);
+		var tapGesture = gestureElement.BindGesture<TGestureElement, TapGestureRecognizer>(commandPath, commandSource, parameterPath, parameterSource);
 
 		return gestureElement.ConfigureTapGesture(tapGesture, numberOfTapsRequired);
 	}
@@ -347,8 +144,8 @@ public static class GesturesExtensions
 	/// <param name="numberOfTapsRequired">Number of taps required to trigger <paramref name="onTapped"/></param>
 	/// <returns><paramref name="gestureElement"/></returns>
 	public static TGestureElement TapGesture<TGestureElement>(this TGestureElement gestureElement,
-																Action onTapped,
-																int? numberOfTapsRequired = null) where TGestureElement : IGestureRecognizers
+		Action onTapped,
+		int? numberOfTapsRequired = null) where TGestureElement : IGestureRecognizers
 	{
 		var gestureRecognizer = new TapGestureRecognizer
 		{
@@ -365,51 +162,10 @@ public static class GesturesExtensions
 		return gestureElement;
 	}
 
-	static Func<TBindingContext, TSource> ConvertExpressionToFunc<TBindingContext, TSource>(in Expression<Func<TBindingContext, TSource>> expression) => expression.Compile();
-
-	static string GetMemberName<T>(in Expression<T> expression) => expression.Body switch
-	{
-		MemberExpression m => m.Member.Name,
-		UnaryExpression { Operand: MemberExpression m } => m.Member.Name,
-		_ => throw new InvalidOperationException("Invalid getter. The `getter` parameter must point directly to a property in the ViewModel and cannot add additional logic")
-	};
-
-	static TGestureRecognizer BindGesture<TGestureRecognizer, TGestureElement, TCommandBindingContext, TParameterBindingContext, TParameterSource>(
-		this TGestureElement gestureElement,
-		Func<TCommandBindingContext, ICommand> getter,
-		(Func<TCommandBindingContext, object?>, string)[] handlers,
-		Action<TCommandBindingContext, ICommand>? setter = null,
-		TCommandBindingContext? source = default,
-		BindingMode commandBindingMode = BindingMode.Default,
-		Func<TParameterBindingContext, TParameterSource>? parameterGetter = null,
-		(Func<TParameterBindingContext, object?>, string)[]? parameterHandlers = null,
-		Action<TParameterBindingContext, TParameterSource>? parameterSetter = null,
-		BindingMode parameterBindingMode = BindingMode.Default,
-		TParameterBindingContext? parameterSource = default)
-		where TGestureElement : IGestureRecognizers
-		where TGestureRecognizer : BindableObject, IGestureRecognizer, new()
-		where TCommandBindingContext : class
-		where TParameterBindingContext : class
-	{
-		var gestureRecognizer = new TGestureRecognizer().BindCommand(getter,
-			handlers,
-			setter,
-			source,
-			commandBindingMode,
-			parameterGetter,
-			parameterHandlers,
-			parameterSetter,
-			parameterSource,
-			parameterBindingMode);
-		gestureElement.GestureRecognizers.Add(gestureRecognizer);
-
-		return gestureRecognizer;
-	}
-
 	static TGestureElement ConfigureSwipeGesture<TGestureElement>(this TGestureElement gestureElement,
-		SwipeGestureRecognizer swipeGesture,
-		SwipeDirection? direction = null,
-		uint? threshold = null) where TGestureElement : IGestureRecognizers
+																	SwipeGestureRecognizer swipeGesture,
+																	SwipeDirection? direction = null,
+																	uint? threshold = null) where TGestureElement : IGestureRecognizers
 	{
 		if (direction is not null)
 		{
@@ -425,8 +181,8 @@ public static class GesturesExtensions
 	}
 
 	static TGestureElement ConfigureTapGesture<TGestureElement>(this TGestureElement gestureElement,
-		TapGestureRecognizer tapGesture,
-		int? numberOfTapsRequired = null) where TGestureElement : IGestureRecognizers
+															TapGestureRecognizer tapGesture,
+															int? numberOfTapsRequired = null) where TGestureElement : IGestureRecognizers
 	{
 		if (numberOfTapsRequired is not null)
 		{
@@ -434,5 +190,20 @@ public static class GesturesExtensions
 		}
 
 		return gestureElement;
+	}
+
+	[RequiresUnreferencedCode("Using bindings with string paths is not trim safe. Use expression-based binding instead.")]
+    static TGestureRecognizer BindGesture<TGestureElement, TGestureRecognizer>(
+		this TGestureElement gestureElement,
+		string commandPath,
+		object? commandSource = null,
+		string? parameterPath = null,
+		object? parameterSource = null) where TGestureElement : IGestureRecognizers
+										where TGestureRecognizer : BindableObject, IGestureRecognizer, new()
+	{
+		var gestureRecognizer = new TGestureRecognizer().BindCommand(commandPath, commandSource, parameterPath, parameterSource);
+		gestureElement.GestureRecognizers.Add(gestureRecognizer);
+
+		return gestureRecognizer;
 	}
 }
